Moves the check of uploaded files with files on the server to the file service
